pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photagger"
version = "1.0.0"
description = "AI-Powered Photography Culling & Tagging Pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Photagger"}]
keywords = ["photography", "ai", "tagging", "lightroom", "xmp", "culling"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "watchdog==4.0.0",
    "PyQt6==6.7.1",
    "opencv-python==4.10.0.84",
    "Pillow==10.4.0",
    "onnxruntime==1.18.1",
    "httpx==0.27.0",
    "numpy==1.26.4",
]

[project.scripts]
photagger = "photagger.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
photagger = ["resources/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
